Spawning / destroying actors, textinput
Can spawn and destroy actors in a specific range from the player, when
in a specific place (blue room in the test level), using the same input.

Started inplementing textinput from the player aswell.

The goal is to connect these two, so that the player can spawn a
keyboard infront of him/her, which can be used for searching files etc
later on in the project. This is not working properly in this version
though, but I'll hopefully get it to work as supposed later today or
tomorrow.