The menu working as it is supposed.
Very much "fulhack", but I'll make it nicer and more general if we will
continue working on this concept.

Regarding the filebrowsing - when running, it gives some errors, which
should be checked prior to the tests. And also, I'm not sure which of
the filebrowsing BP's are the latest, so when clicking the
filebrowser-button in the menu, an old version is loaded for now. Should
of course also be fixed prior to the tests.