Display any imagetype with correct ratios.
Extended the "get texture from file"-class.

Given a path to a picture and a scale (length of the longest side), the
class will load and display the image, with the original ratio kept
intact. No need to specify file-extentions or anything like that. Just
give a path to either a jpg, bmp, png, ico, icns or exr, and the size,
and it should work.